New files: "shell.h" and "shell.c", they are designed to contain all the shell-related functions and variables.
There is a new function, "execute", which allows us to easily run a program (in another process, foreground or background, or in the current one).

The internal functionment is now a bit different, after the parsing of the command line, the shell looks if there is a internal command which matches and executes it. If there is not, the default command
will be executed with the parsed line as parameter.